+++ src/client/main.client.lua
@@ -4,9 +4,12 @@ local StarterPlayerScripts: StarterPlayerScripts = script.Parent
 -- Folders
 local modules: Folder = StarterPlayerScripts:WaitForChild("modules")
 
+-- Random seed
+math.randomseed(tick())
+
 -- Require
 require(modules:WaitForChild("studs_client"))
 require(modules:WaitForChild("clicker_client"))
 
 -- Log
-warn("[Client] All modules loaded successfully!")
+warn("[Main] Startup - All modules loaded successfully")